fixed mutations
Fixed mutations, so there is no longer happening, that generated random
number is the same as the value of gene to be change (hence no change).
Also added a active_genes property to Individual, which outputs the bitmap
indicating, whether gene at given position is part of active node or not.