add evolution method and two gem types
Evolution was encapsulated in single method, with two parameters
objects, in order to simplify usage of CGP. Also two types of
gems were added. First operates on single gene and checks,
whether the gene at given index is the same as in candidate. Second
operates on whole node and checks, whether all genes of mutated node
correspond with candidate.

NO UNIT TESTS in this phase, WIP.